+++ test/EndToEndTests/Tests/Client/Build.Desktop/Microsoft.Test.OData.Tests.Client.NetCore.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <AssemblyName>Microsoft.Test.OData.Tests.Client</AssemblyName>
